+++ examples/auth_ticket.rs
@@ -5,7 +5,7 @@ use std::env::args;
 use std::error::Error;
 use steam_vent::ServerList;
 use steam_vent::auth::{
-    ConsoleAuthConfirmationHandler, DeviceConfirmationHandler, FileGuardDataStore,
+    ClientInfo, ConsoleAuthConfirmationHandler, DeviceConfirmationHandler, FileGuardDataStore,
 };
 use steam_vent::connection::{ReadonlyConnection, UnAuthenticatedConnection};
 use steam_vent_proto::steammessages_clientserver::CMsgClientGameConnectTokens;
@@ -31,6 +31,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
                 ConsoleAuthConfirmationHandler::default(),
                 DeviceConfirmationHandler,
             ),
+            &ClientInfo::default(),
         )
         .await?;