+++ examples/backpack.rs
@@ -2,7 +2,7 @@ use std::env::args;
 use std::error::Error;
 use std::io::Cursor;
 use steam_vent::auth::{
-    ConsoleAuthConfirmationHandler, DeviceConfirmationHandler, FileGuardDataStore,
+    ClientInfo, ConsoleAuthConfirmationHandler, DeviceConfirmationHandler, FileGuardDataStore,
 };
 use steam_vent::{Connection, ConnectionTrait, GenericGCHandshake, ServerList};
 use steam_vent_proto::RpcMessage;
@@ -29,6 +29,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
             ConsoleAuthConfirmationHandler::default(),
             DeviceConfirmationHandler,
         ),
+        &ClientInfo::default(),
     )
     .await?;