+++ examples/inventory.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeMap;
 use std::env::args;
 use std::error::Error;
 use steam_vent::auth::{
-    AuthConfirmationHandler, ConsoleAuthConfirmationHandler, DeviceConfirmationHandler,
+    AuthConfirmationHandler, ClientInfo, ConsoleAuthConfirmationHandler, DeviceConfirmationHandler,
     FileGuardDataStore,
 };
 use steam_vent::proto::csgo::{base_gcmessages::CSOEconItem, gcsdk_gcmessages::CMsgClientHello};
@@ -28,6 +28,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
             ConsoleAuthConfirmationHandler::default(),
             DeviceConfirmationHandler,
         ),
+        &ClientInfo::default(),
     )
     .await?;